added hello world workflows and requirements

+++ khron_brownie/interfaces/token/ERC677Receiver.sol
@@ -2,5 +2,8 @@
 pragma solidity >0.6.0 <0.8.0;
 
 interface ERC677Receiver {
-  function onTokenTransfer(address _sender, uint _value, bytes memory _data) external;
+  function onTokenTransfer(
+    address _sender, 
+    uint256 _value, 
+    bytes memory _data) external;
 }
